refactor(antigravity): address PR #1708 review findings
- Extract shared AntigravityMcp base; AntigravityIdeMcp/AntigravityCliMcp are now thin subclasses (DRY, mirrors antigravity-hooks).

- Extract shared AntigravitySharedSkill base; IDE/CLI skill classes are now thin subclasses.

- MCP: translate canonical url <-> Antigravity serverUrl in both directions so HTTP servers emit the documented Antigravity-compatible shape (was claimed in docs but not implemented).

- Add config-resolver tests for the 'antigravity' alias deprecation warning (once / split-targets / RULESYNC_SILENT_DEPRECATION).

- Docs: note the shared global rule (~/.gemini/GEMINI.md) and hooks (~/.gemini/config/hooks.json) files for the split targets; add Antigravity hook paths to the file-formats note.

- Clarify parsePermissionEntry parsing contract (nested parens round-trip; matches Claude Code).
